+++ openerp/workflow/helpers.py
@@ -1,16 +1,18 @@
+from past.builtins import basestring
+from builtins import object
 import openerp.sql_db
 
 class Session(object):
     def __init__(self, cr, uid):
         assert isinstance(cr, openerp.sql_db.Cursor)
-        assert isinstance(uid, (int, long))
+        assert isinstance(uid, (int, int))
         self.cr = cr
         self.uid = uid
 
 class Record(object):
     def __init__(self, model, record_id):
         assert isinstance(model, basestring)
-        assert isinstance(record_id, (int, long))
+        assert isinstance(record_id, (int, int))
         self.model = model
         self.id = record_id
